에너미가 스턴 당하는거 완?성

+++ Assets/#1 Scripts/#1 Entity/Player/Player.cs
@@ -35,7 +35,7 @@ public class Player : Entity
     /// <returns>
     /// Null
     /// </returns>
-    public override void Setup(float maxHp)
+    public override void Setup(int maxHp)
     {
         base.Setup(maxHp);
